Add utils module
