+++ pyqir-generator/src/lib.rs
@@ -5,3 +5,6 @@
 
 #[cfg(feature = "python-bindings")]
 pub mod python;
+
+#[cfg(feature = "python-bindings")]
+mod utils;
