+++ pyqir-generator/src/python.rs
@@ -9,11 +9,21 @@
 // to resolve it until we move to a newer version of python.
 #![allow(clippy::format_push_string)]
 
+// Lints caused by PyO3 macros:
+#![allow(
+    clippy::borrow_deref_ref,
+    clippy::needless_option_as_deref,
+    clippy::used_underscore_binding
+)]
+
+use crate::utils::{
+    any_to_meta, call_if_some, clone_module, extract_constant, function_type, is_all_same,
+    try_callable_value,
+};
 use pyo3::{
-    exceptions::{PyOSError, PyTypeError, PyUnicodeDecodeError, PyValueError},
+    exceptions::{PyOSError, PyValueError},
     prelude::*,
     types::{PyBytes, PySequence, PyString, PyUnicode},
-    PyClass,
 };
 use qirlib::{
     builder::{BuilderExt, ModuleBuilder},
@@ -21,19 +31,15 @@ use qirlib::{
         builder::Builder as InkwellBuilder,
         context::Context as InkwellContext,
         module::Module as InkwellModule,
-        types::{AnyType, AnyTypeEnum, BasicType, BasicTypeEnum, FunctionType},
-        values::{AnyValue, AnyValueEnum, BasicMetadataValueEnum, CallableValue},
+        types::{AnyType, AnyTypeEnum},
+        values::{AnyValue, AnyValueEnum},
         IntPredicate,
     },
     module,
     qis::BuilderBasicExt,
     types,
 };
-use std::{
-    borrow::Borrow,
-    convert::{Into, TryFrom},
-    mem::transmute,
-};
+use std::{borrow::Borrow, convert::Into, mem::transmute};
 
 #[pymodule]
 fn _native(_py: Python, m: &PyModule) -> PyResult<()> {
@@ -903,16 +909,6 @@ fn bitcode_to_ir<'a>(
     Ok(PyUnicode::new(py, &ir))
 }
 
-fn extract_constant<'ctx>(ty: &impl AnyType<'ctx>, ob: &PyAny) -> PyResult<AnyValueEnum<'ctx>> {
-    match ty.as_any_type_enum() {
-        AnyTypeEnum::IntType(int) => Ok(int.const_int(ob.extract()?, true).into()),
-        AnyTypeEnum::FloatType(float) => Ok(float.const_float(ob.extract()?).into()),
-        _ => Err(PyTypeError::new_err(
-            "Can't convert Python value into this type.",
-        )),
-    }
-}
-
 fn extract_value<'ctx>(ty: &impl AnyType<'ctx>, ob: &PyAny) -> PyResult<AnyValueEnum<'ctx>> {
     match ob.extract::<Value>() {
         Ok(value) => Ok(value.value),
@@ -931,76 +927,6 @@ fn extract_values<'ctx>(
         .collect::<PyResult<Vec<_>>>()
 }
 
-fn function_type<'ctx>(
-    return_type: &impl AnyType<'ctx>,
-    params: impl IntoIterator<Item = AnyTypeEnum<'ctx>>,
-) -> Option<FunctionType<'ctx>> {
-    let params = params
-        .into_iter()
-        .map(|ty| BasicTypeEnum::try_from(ty).map(Into::into).ok())
-        .collect::<Option<Vec<_>>>()?;
-
-    match return_type.as_any_type_enum() {
-        AnyTypeEnum::VoidType(void) => Some(void.fn_type(&params, false)),
-        any => BasicTypeEnum::try_from(any)
-            .map(|basic| basic.fn_type(&params, false))
-            .ok(),
-    }
-}
-
-fn try_callable_value(value: AnyValueEnum) -> Option<(CallableValue, Vec<BasicTypeEnum>)> {
-    match value {
-        AnyValueEnum::FunctionValue(f) => {
-            Some((CallableValue::from(f), f.get_type().get_param_types()))
-        }
-        AnyValueEnum::PointerValue(p) => match p.get_type().get_element_type() {
-            AnyTypeEnum::FunctionType(ty) => {
-                Some((CallableValue::try_from(p).unwrap(), ty.get_param_types()))
-            }
-            _ => None,
-        },
-        _ => None,
-    }
-}
-
-fn any_to_meta(value: AnyValueEnum) -> Option<BasicMetadataValueEnum> {
-    match value {
-        AnyValueEnum::ArrayValue(a) => Some(BasicMetadataValueEnum::ArrayValue(a)),
-        AnyValueEnum::IntValue(i) => Some(BasicMetadataValueEnum::IntValue(i)),
-        AnyValueEnum::FloatValue(f) => Some(BasicMetadataValueEnum::FloatValue(f)),
-        AnyValueEnum::PointerValue(p) => Some(BasicMetadataValueEnum::PointerValue(p)),
-        AnyValueEnum::StructValue(s) => Some(BasicMetadataValueEnum::StructValue(s)),
-        AnyValueEnum::VectorValue(v) => Some(BasicMetadataValueEnum::VectorValue(v)),
-        AnyValueEnum::MetadataValue(m) => Some(BasicMetadataValueEnum::MetadataValue(m)),
-        AnyValueEnum::PhiValue(_)
-        | AnyValueEnum::FunctionValue(_)
-        | AnyValueEnum::InstructionValue(_) => None,
-    }
-}
-
-fn call_if_some(f: Option<&PyAny>) -> PyResult<()> {
-    match f {
-        Some(f) => f.call0().map(|_| ()),
-        None => Ok(()),
-    }
-}
-
-fn is_all_same<T>(py: Python, items: impl IntoIterator<Item = impl Borrow<Py<T>>>) -> bool
-where
-    T: Eq + PyClass,
-{
-    let mut items = items.into_iter();
-    if let Some(mut prev) = items.next() {
-        for item in items {
-            if *item.borrow().borrow(py) != *prev.borrow().borrow(py) {
-                return false;
-            }
-            prev = item;
-        }
-    };
-    true
-}
-
 fn require_same_contexts(
     py: Python,
     contexts: impl IntoIterator<Item = impl Borrow<Py<Context>>>,
@@ -1017,23 +943,4 @@ fn value_contexts<'a>(
 ) -> impl Iterator<Item = Py<Context>> + 'a {
     obs.into_iter()
         .filter_map(|a| Some(a.ok()?.extract::<Value>().ok()?.context))
-}
-
-fn clone_module<'ctx>(
-    module: &InkwellModule,
-    context: &'ctx InkwellContext,
-) -> PyResult<InkwellModule<'ctx>> {
-    let bitcode = module.write_bitcode_to_memory();
-    let new_module = InkwellModule::parse_bitcode_from_buffer(&bitcode, context).map_err(|e| {
-        module.verify().err().map_or_else(
-            || PyOSError::new_err(e.to_string()),
-            |e| PyOSError::new_err(e.to_string()),
-        )
-    })?;
-    let name = module
-        .get_name()
-        .to_str()
-        .map_err(PyUnicodeDecodeError::new_err)?;
-    new_module.set_name(name);
-    Ok(new_module)
 }
